gpu: nvgpu: fifo: fix MISRA 10.3 violations
MISRA Rule 10.3 prohibits assigning objects of different essential or
narrower type. This fixes MISRA 10.3 violations in the common/fifo unit.

JIRA NVGPU-3023

Reviewed-on: https://git-master.nvidia.com/r/2087812
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/gpu/nvgpu/common/fifo/channel_gm20b.c
@@ -55,7 +55,7 @@ void gm20b_channel_bind(struct channel_gk20a *c)
 		 ~ccsr_channel_enable_set_f(~U32(0U))) |
 		 ccsr_channel_enable_set_true_f());
 	nvgpu_smp_wmb();
-	nvgpu_atomic_set(&c->bound, true);
+	nvgpu_atomic_set(&c->bound, (int)true);
 }
 
 u32 gm20b_channel_count(struct gk20a *g)
